LayoutTests/imported/w3c/web-platform-tests/shadow-dom/focus/focus-tabindex-order-shadow-varying-tabindex-2-expected.txt


FAIL Order with different tabindex on host assert_equals: expected Element node <div></div> but got Element node <body>
<script>
promise_test(async () => {
  function cre...
